@@ -1,12 +1,12 @@
 {
 	"config": {
 		"platform": {
-			"php": "7.4"
+			"php": "8.1"
 		}
 	},
 	"require": {
 		"parsecsv/php-parsecsv": "^1.3.2",
-		"ptlis/diff-parser": "^1.0.2"
+		"ptlis/diff-parser": "^1.1.0"
 	},
 	"type": "project"
 }
