@@ -65,7 +65,7 @@ class AttachmentController extends BaseAttachmentController
 		}
 
 		if ($return !== '' && $return !== '0') {
-			$append .= '&return=' . base64_encode($return . ($tmpl ? (strpos($return, '?') === false ? '?' : '&') . 'tmpl=' . $tmpl : ''));
+			$append .= '&return=' . base64_encode($return . ($tmpl ? (str_contains($return, '?') ? '&' : '?') . 'tmpl=' . $tmpl : ''));
 		}
 
 		return $append;
@@ -74,11 +74,11 @@ class AttachmentController extends BaseAttachmentController
 	protected function getReturnPage(): string
 	{
 		$return = $this->input->get('return', null, 'base64');
-		if (empty($return) || !Uri::isInternal(base64_decode($return))) {
+		if (empty($return) || !Uri::isInternal(base64_decode((string)$return))) {
 			return Uri::base();
 		}
 
-		return base64_decode($return);
+		return base64_decode((string)$return);
 	}
 
 	public function getModel($name = 'Form', $prefix = 'Site', $config = ['ignore_request' => true])
