@@ -5,15 +5,15 @@
  * @license    https://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
  */
 
-defined('_JEXEC') or die();
+\defined('_JEXEC') or die();
 
 use Joomla\CMS\Installer\InstallerAdapter;
 use Joomla\CMS\Installer\InstallerScript;
 
 class Com_DPAttachmentsInstallerScript extends InstallerScript
 {
-	protected $minimumPhp      = '7.4.0';
-	protected $minimumJoomla   = '4.0.0';
+	protected $minimumPhp      = '8.1.0';
+	protected $minimumJoomla   = '4.4.0';
 	protected $allowDowngrades = true;
 
 	public function postflight(string $type, InstallerAdapter $parent): void
@@ -22,7 +22,7 @@ class Com_DPAttachmentsInstallerScript extends InstallerScript
 			return;
 		}
 
-		if ($type != 'install' && $type != 'discover_install') {
+		if ($type !== 'install' && $type !== 'discover_install') {
 			return;
 		}
 
