@@ -5,12 +5,12 @@
  * @license    https://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
  */
 
-defined('_JEXEC') or die();
+\defined('_JEXEC') or die();
 
 use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Router\Route;
 
-HTMLHelper::_('script', 'com_dpattachments/dpattachments/views/attachments/default.min.js', ['relative' => true, 'version' => 'auto'], ['defer' => true]);
+HTMLHelper::_('script', 'com_dpattachments/views/attachments/default.min.js', ['relative' => true, 'version' => 'auto'], ['defer' => true, 'type' => 'module']);
 ?>
 <div class="com-dpattachments-attachments">
 	<form action="<?php echo Route::_('index.php?option=com_dpattachments&view=attachments'); ?>"
