@@ -6,7 +6,7 @@
  * @license   https://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
  */
 
-defined('_JEXEC') or die();
+\defined('_JEXEC') or die();
 
 use Joomla\CMS\Factory;
 use Joomla\CMS\HTML\HTMLHelper;
@@ -26,7 +26,7 @@ $app->getLanguage()->load('com_dpattachments', JPATH_ADMINISTRATOR . '/component
 HTMLHelper::_('stylesheet', 'com_dpattachments/dpattachments/layouts/attachments/render.min.css', ['relative' => true]);
 
 HTMLHelper::_('behavior.core');
-HTMLHelper::_('script', 'com_dpattachments/dpattachments/layouts/attachments/render.min.js', ['relative' => true], ['defer' => true]);
+HTMLHelper::_('script', 'com_dpattachments/layouts/attachments/render.min.js', ['relative' => true], ['defer' => true, 'type' => 'module']);
 ?>
 <div class="com-dpattachments-layout-attachments<?php echo $attachments ? '' : ' com-dpattachments-layout-attachments_empty'; ?>">
 	<div class="com-dpattachments-layout-attachments__header"><?php echo Text::_('COM_DPATTACHMENTS_ATTACHMENTS'); ?></div>
