@@ -36,11 +36,11 @@ class HtmlView extends BaseHtmlView
 		if ($app instanceof CMSApplication) {
 			$this->input = $app->input;
 		}
-		$this->item  = $this->get('Item');
-		$this->state = $this->get('State');
+		$this->item  = $this->getModel()->getItem() ?: new \stdClass();
+		$this->state = $this->getModel()->getState();
 
 		// Check for errors.
-		if (count($errors = $this->get('Errors')) > 0) {
+		if (\count($errors = $this->getModel()->getErrors()) > 0) {
 			throw new \Exception(implode('\n', $errors));
 		}
 
@@ -48,7 +48,7 @@ class HtmlView extends BaseHtmlView
 
 		$model = $this->getModel();
 		$model->hit($this->item->id);
-		$this->setLayout(strtolower(pathinfo($this->item->path, PATHINFO_EXTENSION)));
+		$this->setLayout(strtolower(pathinfo((string)$this->item->path, PATHINFO_EXTENSION)));
 
 		require_once JPATH_ADMINISTRATOR . '/components/com_dpattachments/vendor/autoload.php';
 
