@@ -5,7 +5,7 @@
  * @license    https://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
  */
 
-defined('_JEXEC') or die();
+\defined('_JEXEC') or die();
 
 use Joomla\CMS\Factory;
 use Joomla\CMS\HTML\HTMLHelper;
@@ -20,21 +20,21 @@ $changes = $parser->parseFile(Factory::getApplication()->bootComponent('dpattach
 <div class="com-dpattachments-attachment com-dpattachments-attachment-patch">
 	<h3 class="com-dpattachments-attachment__header"><?php echo $this->escape($this->item->title); ?></h3>
 	<div class="com-dpattachments-attachment__content">
-		<?php foreach ($changes->getFiles() as $file) { ?>
+		<?php foreach ($changes->files as $file) { ?>
 			<?php $this->file = $file; ?>
 			<div class="dp-patch-file">
 				<h4 class="dp-patch-file__name"><?php echo $this->loadtemplate('filename'); ?></h4>
-				<?php foreach ($file->getHunks() as $hunk) { ?>
+				<?php foreach ($file->hunks as $hunk) { ?>
 					<table class="dp-table dp-patch-hunk">
-						<?php foreach ($hunk->getLines() as $line) { ?>
-							<tr class="dp-patch-hunk__line dp-patch-hunk__line_<?php echo $line->getOperation(); ?> dp-line">
-								<td class="dp-line__original-nr"><?php echo $line->getOriginalLineNo() != -1 ? $line->getOriginalLineNo() : ''; ?></td>
-								<td class="dp-line__new-nr"><?php echo $line->getNewLineNo() != -1 ? $line->getNewLineNo() : ''; ?></td>
+						<?php foreach ($hunk->lines as $line) { ?>
+							<tr class="dp-patch-hunk__line dp-patch-hunk__line_<?php echo $line->operation; ?> dp-line">
+								<td class="dp-line__original-nr"><?php echo $line->originalLineNo != -1 ? $line->originalLineNo : ''; ?></td>
+								<td class="dp-line__new-nr"><?php echo $line->newLineNo != -1 ? $line->newLineNo : ''; ?></td>
 								<td class="dp-line__operation">
-									<?php echo $line->getOperation() == Line::ADDED ? '+ ' : ''; ?>
-									<?php echo $line->getOperation() == Line::REMOVED ? '- ' : ''; ?>
+									<?php echo $line->operation == Line::ADDED ? '+ ' : ''; ?>
+									<?php echo $line->operation == Line::REMOVED ? '- ' : ''; ?>
 								</td>
-								<td class="dp-line__content"><?php echo htmlentities($line->getContent()); ?></td>
+								<td class="dp-line__content"><?php echo htmlentities($line->content); ?></td>
 							</tr>
 						<?php } ?>
 					</table>
