@@ -5,6 +5,7 @@
  * @license    https://www.gnu.org/licenses/gpl-3.0.html GNU/GPL
  */
 
-defined('_JEXEC') or die();
+\defined('_JEXEC') or die();
 
-echo trim(preg_replace("/\([^)]+\)/", "", $this->file->getOriginalFilename()));
+$file = preg_replace("/\([^)]+\)/", "", (string)$this->file->getOriginalFilename());
+echo trim((string)($file !== '' && $file !== '0'? $file : ''));
