@@ -1,3 +1,4 @@
+
 <?php
 /**
  * @package    DPAttachments
@@ -7,32 +8,23 @@
 
 namespace DigitalPeak\Plugin\Content\DPAttachments\Field;
 
-use DigitalPeak\Component\DPAttachments\Administrator\Extension\DPAttachmentsComponent;
 use Joomla\CMS\Application\CMSWebApplicationInterface;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Form\FormField;
 
 class DpattachmentsField extends FormField
 {
-	public $element;
-	public $form;
 	protected $type = 'Dpattachments';
 
-	protected function getInput()
+	protected function getInput(): string
 	{
 		$app = Factory::getApplication();
-
-		if ((string)$this->element['item_id'] === '' || (string)$this->element['item_id'] === '0') {
-			return $app->getLanguage()->_('PLG_CONTENT_DPATTACHMENTS_FIELD_ATTACHMENTS_NO_ITEM_MESSAGE');
-		}
-
-		$component = $app->bootComponent('dpattachments');
-		if (!$component instanceof DPAttachmentsComponent || !$app instanceof CMSWebApplicationInterface) {
+		if (!$app instanceof CMSWebApplicationInterface) {
 			return '';
 		}
 
 		$app->getDocument()->getWebAssetManager()->addInlineStyle('.com-dpattachments-layout-attachments__header { display: none }');
 
-		return $component->render($this->form->getName(), (string)$this->element['item_id']);
+		return $app->bootComponent('dpattachments')->render($this->form->getName(), (string)$this->element['item_id']);
 	}
 }
